Make function generic over user_data U
